+++ src/main/java/entities/Admin.java
@@ -8,7 +8,7 @@ public class Admin {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private int id;
+    private Long id;
 
     @Column(name = "login")
     private String login;
@@ -16,7 +16,7 @@ public class Admin {
     @Column(name = "password")
     private String password;
 
-    public Admin(int id, String login, String password) {
+    public Admin(Long id, String login, String password) {
         this.id = id;
         this.login = login;
         this.password = password;
@@ -25,11 +25,11 @@ public class Admin {
     public Admin() {
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
